mssql: modify datetime addition test
Truth is: datetime addition in SQL server is not working; adding 5 years
instead of 10 doesn't cause the bug we used to encounter in the CI.

Even though the number of seconds is correctly passed to the SQL query,
SQL server might be having some overflowing somewhere.

Even when using 5 years instead of 10:

delta_1 = @updated_at - durPos
delta_2 = @updated_at + durPos

We would expect delta_1 = -delta_2, but it's not.