ci: run the commands directly
There are some weird failures when trying to install in certain env and
certain tests, which doesn't make any sense. why would a certain config
install for certain tests but not others. And why jruby+mssql is
screwing around?

I tested the failures locally exactly as the the things run in the CI
but I couldn't reproduce. weird.